feat: Error handler
