+++ tests/StacktraceTest.php
@@ -12,26 +12,25 @@
 namespace Sentry\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Sentry\ClientBuilder;
-use Sentry\ClientInterface;
 use Sentry\Frame;
+use Sentry\Options;
 use Sentry\Stacktrace;
 
 class StacktraceTest extends TestCase
 {
     /**
-     * @var ClientInterface
+     * @var Options
      */
-    protected $client;
+    protected $options;
 
     protected function setUp()
     {
-        $this->client = ClientBuilder::create()->getClient();
+        $this->options = new Options();
     }
 
     public function testGetFramesAndToArray()
     {
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
 
         $stacktrace->addFrame('path/to/file', 1, ['file' => 'path/to/file', 'line' => 1, 'function' => 'test_function']);
         $stacktrace->addFrame('path/to/file', 2, ['file' => 'path/to/file', 'line' => 2, 'function' => 'test_function', 'class' => 'TestClass']);
@@ -46,7 +45,7 @@ class StacktraceTest extends TestCase
 
     public function testStacktraceJsonSerialization()
     {
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
 
         $stacktrace->addFrame('path/to/file', 1, ['file' => 'path/to/file', 'line' => 1, 'function' => 'test_function']);
         $stacktrace->addFrame('path/to/file', 2, ['file' => 'path/to/file', 'line' => 2, 'function' => 'test_function', 'class' => 'TestClass']);
@@ -61,7 +60,7 @@ class StacktraceTest extends TestCase
 
     public function testAddFrame()
     {
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
         $frames = [
             $this->getJsonFixture('frames/eval.json'),
             $this->getJsonFixture('frames/runtime_created.json'),
@@ -82,7 +81,7 @@ class StacktraceTest extends TestCase
 
     public function testAddFrameSerializesMethodArguments()
     {
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
         $stacktrace->addFrame('path/to/file', 12, [
             'file' => 'path/to/file',
             'line' => 12,
@@ -99,9 +98,9 @@ class StacktraceTest extends TestCase
 
     public function testAddFrameStripsPath()
     {
-        $this->client->getOptions()->setPrefixes(['path/to/', 'path/to/app']);
+        $this->options->setPrefixes(['path/to/', 'path/to/app']);
 
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
 
         $stacktrace->addFrame('path/to/app/file', 12, ['function' => 'test_function_parent_parent_parent']);
         $stacktrace->addFrame('path/to/file', 12, ['function' => 'test_function_parent_parent']);
@@ -118,10 +117,10 @@ class StacktraceTest extends TestCase
 
     public function testAddFrameMarksAsInApp()
     {
-        $this->client->getOptions()->setProjectRoot('path/to');
-        $this->client->getOptions()->setExcludedProjectPaths(['path/to/excluded/path']);
+        $this->options->setProjectRoot('path/to');
+        $this->options->setExcludedProjectPaths(['path/to/excluded/path']);
 
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
 
         $stacktrace->addFrame('path/to/file', 12, ['function' => 'test_function']);
         $stacktrace->addFrame('path/to/excluded/path/to/file', 12, ['function' => 'test_function']);
@@ -135,7 +134,7 @@ class StacktraceTest extends TestCase
     public function testAddFrameReadsCodeFromShortFile()
     {
         $fileContent = explode("\n", $this->getFixture('code/ShortFile.php'));
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
 
         $stacktrace->addFrame($this->getFixturePath('code/ShortFile.php'), 3, ['function' => '[unknown]']);
 
@@ -159,7 +158,7 @@ class StacktraceTest extends TestCase
     public function testAddFrameReadsCodeFromLongFile()
     {
         $fileContent = explode("\n", $this->getFixture('code/LongFile.php'));
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
 
         $stacktrace->addFrame($this->getFixturePath('code/LongFile.php'), 8, [
             'function' => '[unknown]',
@@ -192,7 +191,7 @@ class StacktraceTest extends TestCase
             $this->expectExceptionMessage('Invalid frame index to remove.');
         }
 
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
 
         $stacktrace->addFrame('path/to/file', 12, [
             'function' => 'test_function_parent',
@@ -224,7 +223,7 @@ class StacktraceTest extends TestCase
     public function testFromBacktrace()
     {
         $fixture = $this->getJsonFixture('backtraces/exception.json');
-        $frames = Stacktrace::createFromBacktrace($this->client, $fixture['backtrace'], $fixture['file'], $fixture['line'])->getFrames();
+        $frames = Stacktrace::createFromBacktrace($this->options, $fixture['backtrace'], $fixture['file'], $fixture['line'])->getFrames();
 
         $this->assertFrameEquals($frames[0], null, 'path/to/file', 16);
         $this->assertFrameEquals($frames[1], 'TestClass::crashyFunction', 'path/to/file', 7);
@@ -234,7 +233,7 @@ class StacktraceTest extends TestCase
     public function testFromBacktraceWithAnonymousFrame()
     {
         $fixture = $this->getJsonFixture('backtraces/anonymous_frame.json');
-        $frames = Stacktrace::createFromBacktrace($this->client, $fixture['backtrace'], $fixture['file'], $fixture['line'])->getFrames();
+        $frames = Stacktrace::createFromBacktrace($this->options, $fixture['backtrace'], $fixture['file'], $fixture['line'])->getFrames();
 
         $this->assertFrameEquals($frames[0], null, 'path/to/file', 7);
         $this->assertFrameEquals($frames[1], 'call_user_func', '[internal]', 0);
@@ -250,7 +249,7 @@ class StacktraceTest extends TestCase
             null,
         ];
 
-        $stacktrace = new Stacktrace($this->client);
+        $stacktrace = new Stacktrace($this->options);
         $stacktrace->addFrame('/some/file', 123, $stack);
         $frames = $stacktrace->getFrames();
